pyproject.toml
[project]
name = "backtrader-research"
version = "0.3.0"
description = "Autonomous trading strategy research"
requires-python = ">=3.10"
dependencies = [
    "yfinance>=0.2.36",
    "backtrader>=1.9.78",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["README.md"]
